Add basic functionality of leapp report analyzer
Introduce script to analyze and group multiple leapp report JSON files
into a single JSON output and an HTML tables output.

The scripts takes a list of input files, merges their data, and groups
the hostnames by severity. Output is sorted by severity.

Add support for checking the 'groups' field if the 'flags' field is not
present in the leapp report entry. So we can parse multiple versions of
leapp report schemas.

The script seems to work for both python2.7 and 3+ versions.